make dockerfile multi arch compatible
In order to make the dockerfile handle multi arch builds, I mostly removed the
build target. Then, I optimized the dockerfile by splitting the fetch step from
the build step. By doing so, only when we change a Cargo.toml or Cargo.lock file,
the fetch step will be re-executed. And by adding --platform to that step, it will
only be executed once for all platforms and with the current platform target,
making it more efficient.
Then I also gave a name to the build steps in improve readability.
Finally, I updated the Makefile so that we can build and push with buildx.

Signed-off-by: Jérémie Drouet <jeremie.drouet@gmail.com>